refactor(core): delete the factory's ASGI assembly, which nothing served
`create_health_routes`, `create_combined_asgi_app`, `create_auth_combined_app`
and `MCPServerFactory.create_asgi_app` had one production caller between them:
each other. `serve --http` builds its app in
`server/lifecycle.mcp_app_for_serving` and wraps it with
`create_auth_enforced_app`.

The two assemblies had drifted, and that is the finding rather than a detail:
the deleted routes were flat `/health` and `/ready`, while a running Hangar
serves `/health/live`, `/health/ready`, `/health/startup`. The tests driving
this path were asserting on routes production does not have.

After the cut the factory builds no HTTP surface at all -- only
`create_server()`. That collapses two guards that existed because it did:

  * `test_trusted_hosts_reach_the_mcp_endpoint` parametrised the transport-
    security check over both construction paths, on the stated grounds that
    this repo keeps shipping a capability wired into one and not the other.
    Narrowed to the one path that remains, with the docstring saying why and
    telling the next person to add a leg back if a second path returns.
    Verified the surviving leg still bites: setting `transport_security=None`
    in `lifecycle.py` fails it.
  * `test_the_factory_surface_shares_the_router` asserted the factory reached
    approvals through `create_api_router` "so the two cannot diverge". With no
    surface left, there is nothing to diverge; deleted rather than weakened
    into a tautology.

`test_asgi_identity_bridge.py` is split rather than deleted. Half of it tested
`_principal_to_identity_context`, which is live and out of scope; half drove the
dying combiner. The live path does not bind identity in an ASGI wrapper at all
-- since #576 `mcp_tool_wrapper` sets `identity_context_var` before the tool
body runs (`server/tools/mcp_server.py` says so). The mapping tests survive
under a name that does not point at a deleted function.

Also gone with their subject: `TestCombinedAppRouting`, `TestAuthCombinedAppRouting`
(`TestApiRouterWsMounts` stays -- `create_api_router` is live) and
`TestASGIIntegration`, whose `/health` and `/api/` mounting is covered on the
served app by `tests/integration/test_rest_authz_on_served_app.py` and
`tests/live/test_t0_smoke.py`.

Left alone per #953: `bind_caller_identity`, `release_caller_identity`,
`mcp_transport_security`, and the unwired `_ws_handshake_allowed` /
`_strip_host_port` security helpers, which this PR leaves with no `src/` caller
by design.

Closes #955
Refs #953
