+++ tests/unit/test_api_providers.py
@@ -352,64 +352,3 @@ class TestGetProviderHealth:
 # ---------------------------------------------------------------------------
 # ASGI integration: API mount + existing routes
 # ---------------------------------------------------------------------------
-
-
-class TestASGIIntegration:
-    """Tests verifying the API is mounted at /api/ and existing routes still work."""
-
-    @pytest.fixture
-    def full_app_client(self, mock_context):
-        """TestClient for the full combined ASGI app."""
-        from unittest.mock import Mock
-
-        from mcp_hangar.fastmcp_server.factory import MCPServerFactory
-        from mcp_hangar.fastmcp_server.config import HangarFunctions
-
-        hangar = HangarFunctions(
-            list=Mock(return_value={"mcp_servers": []}),
-            start=Mock(return_value={"status": "started"}),
-            stop=Mock(return_value={"status": "stopped"}),
-            invoke=Mock(return_value={"result": 42}),
-            tools=Mock(return_value={"tools": []}),
-            details=Mock(return_value={"mcp_server": "test"}),
-            health=Mock(return_value={"status": "healthy"}),
-        )
-
-        factory = MCPServerFactory(hangar)
-
-        with patch("mcp_hangar.server.api.middleware.get_context", return_value=mock_context):
-            with patch("mcp_hangar.fastmcp_server.factory.MCPServerFactory.create_server") as mock_server:
-                # Avoid actually starting server in test
-                from starlette.responses import JSONResponse
-
-                async def mcp_endpoint(scope, receive, send):
-                    response = JSONResponse({"mcp": "ok"})
-                    await response(scope, receive, send)
-
-                mock_server.return_value = Mock()
-                mock_server.return_value.streamable_http_app.return_value = mcp_endpoint
-
-                app = factory.create_asgi_app()
-                yield TestClient(app, raise_server_exceptions=False)
-
-    def test_health_endpoint_still_works(self, full_app_client):
-        """GET /health still returns ok response."""
-        response = full_app_client.get("/health")
-        assert response.status_code == 200
-        data = response.json()
-        assert data["status"] == "ok"
-
-    def test_api_providers_reachable_under_api_prefix(self, full_app_client):
-        """GET /api/mcp_servers is reachable via the /api prefix."""
-        response = full_app_client.get("/api/mcp_servers/")
-        assert response.status_code == 200
-
-    def test_ready_endpoint_still_works(self, full_app_client):
-        """GET /ready still works after adding API mount."""
-        response = full_app_client.get("/ready")
-        assert response.status_code in (200, 503)  # 503 when checks fail in test env
-
-    def test_metrics_endpoint_still_works(self, full_app_client):
-        """GET /metrics still works after adding API mount."""
-        response = full_app_client.get("/metrics")
-        assert response.status_code == 200
