+++ tests/unit/test_ws_routing.py
@@ -8,7 +8,7 @@ Tests verify:
 - create_api_router includes /ws mount with events and state routes
 """
 
-from unittest.mock import AsyncMock, MagicMock, patch
+from unittest.mock import patch
 
 from starlette.routing import Mount
 
@@ -17,151 +17,14 @@ from starlette.routing import Mount
 # ---------------------------------------------------------------------------
 
 
-def _http_scope(path: str) -> dict:
-    return {"type": "http", "path": path, "method": "GET", "headers": [], "client": None}
-
-
-def _ws_scope(path: str) -> dict:
-    return {"type": "websocket", "path": path, "headers": [], "client": None}
-
-
-def _lifespan_scope() -> dict:
-    return {"type": "lifespan"}
-
-
 # ---------------------------------------------------------------------------
 # combined_app routing tests
 # ---------------------------------------------------------------------------
 
 
-class TestCombinedAppRouting:
-    """Test create_combined_asgi_app websocket + http routing."""
-
-    def _make_combined_app(self):
-        from mcp_hangar.fastmcp_server.asgi import create_combined_asgi_app
-
-        aux_app = AsyncMock()
-        mcp_app = AsyncMock()
-        api_app = AsyncMock()
-        app = create_combined_asgi_app(aux_app, mcp_app, api_app)
-        return app, aux_app, mcp_app, api_app
-
-    async def test_websocket_api_ws_events_goes_to_api_app(self):
-        """websocket scope on /api/ws/events is forwarded to api_app."""
-        app, aux_app, mcp_app, api_app = self._make_combined_app()
-        scope = _ws_scope("/api/ws/events")
-
-        await app(scope, AsyncMock(), AsyncMock())
-
-        api_app.assert_called_once()
-        mcp_app.assert_not_called()
-
-    async def test_websocket_api_ws_state_goes_to_api_app(self):
-        """websocket scope on /api/ws/state is forwarded to api_app."""
-        app, aux_app, mcp_app, api_app = self._make_combined_app()
-        scope = _ws_scope("/api/ws/state")
-
-        await app(scope, AsyncMock(), AsyncMock())
-
-        api_app.assert_called_once()
-        mcp_app.assert_not_called()
-
-    async def test_websocket_non_api_path_goes_to_mcp_app(self):
-        """websocket scope on /mcp is NOT forwarded to api_app."""
-        app, aux_app, mcp_app, api_app = self._make_combined_app()
-        scope = _ws_scope("/mcp")
-
-        await app(scope, AsyncMock(), AsyncMock())
-
-        mcp_app.assert_called_once()
-        api_app.assert_not_called()
-
-    async def test_http_health_still_goes_to_aux_app(self):
-        """Regression: http scope on /health still goes to aux_app."""
-        app, aux_app, mcp_app, api_app = self._make_combined_app()
-        scope = _http_scope("/health")
-
-        await app(scope, AsyncMock(), AsyncMock())
-
-        aux_app.assert_called_once()
-        mcp_app.assert_not_called()
-        api_app.assert_not_called()
-
-    async def test_http_api_providers_still_goes_to_api_app(self):
-        """Regression: http scope on /api/mcp_servers/ still goes to api_app."""
-        app, aux_app, mcp_app, api_app = self._make_combined_app()
-        scope = _http_scope("/api/mcp_servers/")
-
-        await app(scope, AsyncMock(), AsyncMock())
-
-        api_app.assert_called_once()
-        mcp_app.assert_not_called()
-
-    async def test_websocket_api_prefix_stripped_before_forwarding(self):
-        """WebSocket scope path has /api prefix stripped: /api/ws/events -> /ws/events."""
-        app, aux_app, mcp_app, api_app = self._make_combined_app()
-        scope = _ws_scope("/api/ws/events")
-
-        captured_scopes = []
-
-        async def capturing_api_app(scope, receive, send):
-            captured_scopes.append(dict(scope))
-
-        from mcp_hangar.fastmcp_server.asgi import create_combined_asgi_app
-
-        app2 = create_combined_asgi_app(aux_app, mcp_app, capturing_api_app)
-
-        await app2(scope, AsyncMock(), AsyncMock())
-
-        assert len(captured_scopes) == 1
-        assert captured_scopes[0]["path"] == "/ws/events"
-
-
 # ---------------------------------------------------------------------------
 # auth_combined_app routing tests
 # ---------------------------------------------------------------------------
-
-
-class TestAuthCombinedAppRouting:
-    """Test create_auth_combined_app websocket routing."""
-
-    def _make_auth_app(self):
-        from mcp_hangar.fastmcp_server.asgi import create_auth_combined_app
-
-        aux_app = AsyncMock()
-        mcp_app = AsyncMock()
-        api_app = AsyncMock()
-
-        # Minimal stubs for auth_components and config
-        auth_components = MagicMock()
-        auth_components.authn_middleware.authenticate.return_value = {"user": "test"}
-
-        config = MagicMock()
-        config.auth_skip_paths = ["/health", "/ready", "/metrics"]
-        config.trusted_proxies = []
-
-        app = create_auth_combined_app(aux_app, mcp_app, auth_components, config, api_app)
-        return app, aux_app, mcp_app, api_app
-
-    async def test_auth_app_websocket_api_ws_events_goes_to_api_app(self):
-        """auth_combined_app: websocket on /api/ws/events goes to api_app (no auth)."""
-        app, aux_app, mcp_app, api_app = self._make_auth_app()
-        scope = _ws_scope("/api/ws/events")
-
-        await app(scope, AsyncMock(), AsyncMock())
-
-        api_app.assert_called_once()
-        mcp_app.assert_not_called()
-
-    async def test_auth_app_lifespan_falls_through_to_mcp_app(self):
-        """auth_combined_app: lifespan scope goes directly to mcp_app."""
-        app, aux_app, mcp_app, api_app = self._make_auth_app()
-        scope = _lifespan_scope()
-
-        await app(scope, AsyncMock(), AsyncMock())
-
-        mcp_app.assert_called_once()
-        api_app.assert_not_called()
 
 
 # ---------------------------------------------------------------------------
